requirements.txt
pandas
vyper-config
